refactor(chat): extract shared ACP host client for CLI pools
Share createAcpHostClient and acquireAcpCliConnection between providers.
Refactor Cursor ACP pool to delegate spawn/init and cursor_login handling.
